requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
defusedxml==0.7.1
idna==3.10
jira==3.10.5
oauthlib==3.3.1
packaging==25.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
six==1.17.0
typing_extensions==4.15.0
urllib3==2.5.0
